Adds AWS authentication module
Introduces a PowerShell module that streamlines authentication with AWS services like CodeArtifact and ECR.

This module includes functions for:
- Retrieving the AWS region
- Obtaining the domain owner ID
- Getting the CodeArtifact domain name
- Ensuring the AWS SSO token is valid, prompting the user to re-authenticate if necessary

These functions are used by the `refreshNugetToken.ps1` and `refreshEcrDockerToken.ps1` scripts. The module handles AWS SSO login and token validation, improving the user experience and security. Additionally, the Readme file is updated to reflect these changes and to explain how to integrate SSO.